refactor: move classes that talk to UserPreference table to the user-preference folder
