+++ src/main/factories/infra/db/prisma/user/add-user-preference-prisma-repo-factory.ts
@@ -1,4 +1,4 @@
-import { AddUserPreferencePrismaRepo } from '@/infra/db/prisma/data/user/add-user-preference/add-user-preference-prisma-repo'
+import { AddUserPreferencePrismaRepo } from '@/infra/db/prisma/data/user-preference/add-user-preference/add-user-preference-prisma-repo'
 import { type AddUserPreferenceRepo } from '@/usecases/contracts/db/user'
 
 export const makeAddUserPreferencePrismaRepo = (): AddUserPreferenceRepo => {
