+++ src/main/factories/infra/db/prisma/user/find-preference-by-id-prisma-repo-factory.ts
@@ -1,4 +1,4 @@
-import { FindPreferenceByIdPrismaRepo } from '@/infra/db/prisma/data/user/find-preference-by-id/find-preference-by-id-prisma-repo'
+import { FindPreferenceByIdPrismaRepo } from '@/infra/db/prisma/data/user-preference/find-preference-by-id/find-preference-by-id-prisma-repo'
 import { type FindPreferenceByIdRepo } from '@/usecases/contracts/db/user'
 
 export const makeFindPreferenceByIdPrismaRepo = (): FindPreferenceByIdRepo => {
