+++ src/main/factories/infra/db/prisma/user/update-user-preference-repo-factory.ts
@@ -1,4 +1,4 @@
-import { UpdateUserPreferencePrismaRepo } from '@/infra/db/prisma/data/user/update-user-preferences/update-user-preference-prisma-repo'
+import { UpdateUserPreferencePrismaRepo } from '@/infra/db/prisma/data/user-preference/update-user-preference/update-user-preference-prisma-repo'
 import { type UpdateUserPreferenceRepo } from '@/usecases/contracts/db/user'
 
 export const makeUpdateUserPreferenceRepo = (): UpdateUserPreferenceRepo => {
